refactor: restructure utilities
