+++ src/routes/volumes/[volumeName]/+page.svelte
@@ -9,7 +9,7 @@
 	import * as Alert from '$lib/components/ui/alert/index.js';
 	import { Separator } from '$lib/components/ui/separator/index.js';
 	import { Badge } from '$lib/components/ui/badge/index.js';
-	import { formatDate } from '$lib/utils';
+	import { formatDate } from '$lib/utils/string.utils';
 
 	let { data, form }: { data: PageData; form: ActionData } = $props();
 	let { volume, inUse } = $derived(data);
