+++ tailwind.config.ts
@@ -1,8 +1,9 @@
-import { fontFamily } from 'tailwindcss/defaultTheme';
+import fontFamily from 'tailwindcss/defaultTheme';
 import type { Config } from 'tailwindcss';
+import tailwindcssAnimate from 'tailwindcss-animate';
 
 const config: Config = {
-	darkMode: ['class'],
+	darkMode: 'class',
 	content: ['./src/**/*.{html,js,svelte,ts}'],
 	safelist: ['dark'],
 	theme: {
